Separated date criteria view into an elements and set default criteria elements
